move InputNameWarning
